requirements.txt
# Requirements
azure-common==1.1.23
azure-mgmt-compute==12.0.0
azure-mgmt-monitor==0.9.0
azure-mgmt-resource==9.0.0
azure-mgmt-subscription==0.5.0
numpy==1.18.3
# csv
